requirements.txt
openai==1.42.0
pydantic==2.8.2
pydantic_settings==2.4.0
python-dotenv==1.0.1
fastapi==0.112.2
uvicorn==0.30.6
python-multipart==0.0.9
requests==2.32.3